:art: Move mechanism and policy logic to modules
