+++ model/policies/__init__.py
@@ -1,0 +1,7 @@
+from .claim import claim_policy
+from .claim_mTokens import claimMTokens_policy
+from .deposit import deposit_policy
+from .exchange import exchange_policy
+from .liquidate import liquidate_policy, vote_liquidate_policy
+from .refund import refund_policy
+
